--- examples/cmd/whoami/main.go
@@ -1,58 +1,0 @@
-package main
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"os"
-
-	pineconego "github.com/henomis/pinecone-go"
-	"github.com/henomis/pinecone-go/request"
-	"github.com/henomis/pinecone-go/response"
-)
-
-func main() {
-
-	apiKey := os.Getenv("PINECONE_API_KEY")
-	if apiKey == "" {
-		panic("PINECONE_API_KEY is not set")
-	}
-
-	environment := os.Getenv("PINECONE_ENVIRONMENT")
-	if environment == "" {
-		panic("PINECONE_ENVIRONMENT is not set")
-	}
-
-	p := pineconego.New(environment, apiKey)
-
-	req := &request.Whoami{}
-	res := &response.Whoami{}
-	err := p.Whoami(context.Background(), req, res)
-	if err != nil {
-		panic(err)
-	}
-
-	if !res.IsSuccess() {
-		if res.Response.Code != nil {
-			fmt.Printf("Error: %d -", *res.Response.Code)
-		}
-
-		if res.Response.Message != nil {
-			fmt.Printf(" %s\n", *res.Response.Message)
-		}
-
-		if res.Response.RawBody != nil {
-			fmt.Printf(" %s\n", *res.Response.RawBody)
-		}
-
-		return
-	}
-
-	b, err := json.MarshalIndent(res, "", "  ")
-	if err != nil {
-		panic(err)
-	}
-
-	fmt.Println(string(b))
-
-}
